requirements.txt
plotly.express
pandas
folium
geopy
geopy.geocoders
Nominatum
uszipcode